moved root file to server folder
