+++ server/test/repayment.spec.js
@@ -1,6 +1,6 @@
 import chai, { assert } from 'chai';
 import chaiHttp from 'chai-http';
-import server from '../../index';
+import server from '../index';
 import userInfo from './userInfo';
 import loanInfo from './loanInfo';
